Move ChainOpts to ETH chain service
